Preserve trailing slash in URL path in runtime
Given a path like "/api/tasks/" the runtime uses `path.Join` to join the
base path to it which strips any trailing slash. If the path previously
ended with a trailing slash ("/"), it is reinstated after adding the
base path prefix because the behaviour of the server's path handling
cannot be guaranteed.

Fixes #289

Signed-off-by: Jonathan Ingram <jonathan.b.ingram@gmail.com>