Make File implement io.Reader
When go-swagger generated an api with an endpoint that returned a `file`
with the content-type of `application/octet-stream` the endpoint would
panic. The `application/octet-stream `maps to
`ByteStream{Producer|Consumer}` and the `ByteStreamProducer` can only
serialize implementations of `io.Reader`. This Fixes #42, and adds a
simple test that asserts that `File` implements io.Reader