Add additional context to ReadResponse errors
When there is an issue reading a response, because the API is sending an unexpected content type for example, the error that ends up bubbling up is often something that isn’t very useful. e.g.

``&{0 } (*models.Error) is not supported by the TextConsumer, can be resolved by supporting TextUnmarshaler interfac``

This wraps the errors that come out of readResponse.ReadResponse and add a few pieces of key information that can provide more context to what’s going on, like the statusCode and the Content-Type.

Signed-off-by: Bret Comnes <bcomnes@gmail.com>