Bugfix: Enforce Accept header validation
Validation of the `Accept` header was only being enforced when the
incoming request includes a body.

Removes the check for the request having a body enforces the `Accept`
header is valid.

Removes the default response format of `*/*` when no body is included
in the request.

Also includes basic lint fixes by using `gofmt`, `goimports`.

resolves: #172
Signed-off-by: kenjones <kenjones@cisco.com>