Bugfix: Handle Accept header not provided
As part of #172, the Accept header validation was being enforced.
This resulted in a backwards-compatibility issues where if Accept
header was not provided, it would result in an error.
When an Accept header is not provided the client is indicating it
will accept the default response format. As such it should not result
in an error.

Applies fix to provide a default value when no default could be
identified from the request body or using the defined produces on the
specific route.

Applies fix to validation to consider the request valid if no format
could be identified using just the request and produces is empty.

resolves: #174
Signed-off-by: kenjones <kenjones@cisco.com>