When response body is binary, DEBUG mode should not print it
When the response is a binary blob, printing it can mess up the
user's terminal. This change checks the response type and, if
the response is an octet-stream, omits the response body from
the logs.

Note that octet-stream is not the default response body type, so
it is unlikely that we will falsely omit a response body which
was not actually binary.

Signed-off-by: andrew2nelson <andrew2nelson@arista.com>