Allows base paths to include query parameters
This is a test case and a fix for https://github.com/go-openapi/runtime/issues/207.

When the base path of an operation contains a query string, those parameters are:

1. extracted before appending any path parameters to the path
2. merged with the query parameters

In the case of a conflict between the parameters passed in the API call and the ones declared in the base path, the former prevails. This gives the end control to clients.

Signed-off-by: Julien Silland <julien@soliton.io>