Allow specifying a content type
Currently, the content type header for a multipart file is
automatically detected with `http.DetectContentType`.
However, sometimes we need to specify a custom content type.

This commit provides a simple way to pass a custom content type.
If the given `NamedReadCloser` also has a function
`ContentType() string`, the returned value will be used.

Signed-off-by: Junpei Kawamoto <kawamoto.junpei@gmail.com>