chore(deps): deprecated dependency to opentracing
* fixes #365

This PR deprecates the WithOpenTracing runtime option,
in favor the WithOpenTelemetry option.

The WithOpenTracing method remains exposed, so programs that build
with that one don't break, but options are ignored and it returns
an opentelemetry-enabled transport.

A new independent is introduced to provide backward-compatibility to
users who are not able to transition to opentelemetry.

Technically, the runtime repo now moves to a multi-modules mono-repo
structure, similar to github.com/go-openapi/swag. The necessary CI
changes are applied.

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>